ovs-router: Add infrastructure for multi-table routing.
For route lookup based on the destination address a single routing table
is sufficient. For a more advanced routing when a lookup needs to take
into consideration other parameters, such as a source address, a
multi-table lookup is needed.

This change introduces infrastructure for using multiple routing
tables that can be added dynamically, as a pre-step towards importing
non-default routing tables from kernel.

Signed-off-by: Dima Chumak <dchumak@nvidia.com>